+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to **Dangl.Data.Shared** are documented here.
 
 ## v1.1.0:
 - The `Dangl.Data.Shared.AspNetCore` package was updated to ASP.NET Core 2.2
+- Add `LocalizedSpaStaticFileExtensions` to help serving localized Single Page Applications (SPAs)
+- Add `IUserLanguageService` with default implementation
 
 ## v1.0.10:
 - Add `CompressedRequestMiddleware` to support clients sending compressed request bodies